separate Action into own file and parsing
Introducing the `on.fail` field for the kube gdt plugin involves parsing
a kubernetes action (get, create, apply, etc) separate from the main
`kube` field in the KubeSpec. So, in this patch we separate out the
Action and then embed the Action struct inside the KubeSpec, separating
out the parsing of the Action struct.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>