allow KinD clusters to be deleted or retained
Modifies the KinD Fixture to make it more stable and more flexible.

The default behaviour is for the KinD Fixture to delete the KinD cluster
only if the KinD cluster existed before the Fixture is started.
Similarly, if the KinD Cluster did *not* exist before the Fixture is
started, the Fixture deletes the KinD Cluster when it is stopped.

This commit adds a WithDeleteOnStop() and WithRetainOnStop() modifier to
the KinD Fixture creation allowing users to indicate if they want to
change that default behaviour.

In addition to the above enhancements, this commit makes our testing
more stable by adding some KinD cluster cleanup to our Makefile as well
as serializing execution of longer-running, resource-intensive tests.

Issue #13

Signed-off-by: Jay Pipes <jaypipes@gmail.com>