fix bug in var jsonpath selection
When looking up a JSONPath in the output of a list operation, one
specifies the path like so: `$[0].status.podIP`. However, the way that
we were transforming UnstructuredList to any was incorrect and therefore
the github.com/theory/jsonpath.Select() function wasn't understanding
the provided input. This patch fixes this problem by changing the input
from []map[string]any to []any.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>