support `assert.require` to fail test scenario
Adds support for the `assert.require` field which contains a boolean
value that instructs the gdt scenario to exit with a failure when an
assertion fails. This is equivalent to the difference between testify
assert and testify require or the Go testing.T.Error versus
testing.T.Fatal.

Issue gdt-dev/gdt#66

Signed-off-by: Jay Pipes <jaypipes@gmail.com>